Move visdom.
